sécurité

+++ backend/controllers/sauce.js
@@ -1,4 +1,4 @@
-// 
+const jwt = require('jsonwebtoken');
 
 const Sauce = require('../models/sauce');
 const fs = require('fs');
@@ -39,21 +39,40 @@ exports.modifySauce = (req, res, next) => {
       ...JSON.parse(req.body.sauce),
       imageUrl: `${req.protocol}://${req.get('host')}/images/${req.file.filename}`
     } : { ...req.body };
+    //tester si l'auteur est bien le meme que celui qui fait la requete.
+
+    const token = req.headers.authorization.split(' ')[1];
+    const decodedToken = jwt.verify(token, 'RANDOM_TOKEN_SECRET');
+    const userId = decodedToken.userId;
+    if (userId !== sauceObject.userId){
+      res.status(403).json({ message : "seul l'auteur peut faire cette requete"});
+    } else{
   Sauce.updateOne({ _id: req.params.id }, { ...sauceObject, _id: req.params.id })
     .then(() => res.status(200).json({ message: 'Objet modifié !'}))
     .catch(error => res.status(400).json({ error }));
+    }
 };
 
 exports.deleteSauce = (req, res, next) => {
   Sauce.findOne({ _id: req.params.id })
     .then(sauce => {
+      //seul l'auteur d'une sauce peut supprimer celle-ci
+    const token = req.headers.authorization.split(' ')[1];
+    const decodedToken = jwt.verify(token, 'RANDOM_TOKEN_SECRET');
+    const userId = decodedToken.userId;
+    if (userId !== sauce.userId){
+      res.status(403).json({ message : "seul l'auteur d'une sauce peut supprimer celle-ci"});
+
+    } else {
       const filename = sauce.imageUrl.split('/images/')[1];
       fs.unlink(`images/${filename}`, () => {
         Sauce.deleteOne({ _id: req.params.id })
           .then(() => res.status(200).json({ message: 'Objet supprimé !'}))
           .catch(error => res.status(400).json({ error }));
       });
+    }
     })
+  
     .catch(error => res.status(500).json({ error }));
 };
 exports.getAllSauce = (req, res, next) => {
